internal/dag: split Builder and KubernetesCache types
