+++ internal/dag/builder.go
@@ -20,122 +20,21 @@ import (
 	"sort"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/api/extensions/v1beta1"
 	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/client-go/tools/cache"
 
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/auth"
 	ingressroutev1 "github.com/heptio/contour/apis/contour/v1beta1"
 )
-
-// A KubernetesCache holds Kubernetes objects and associated configuration and produces
-// DAG values.
-type KubernetesCache struct {
-	// IngressRouteRootNamespaces specifies the namespaces where root
-	// IngressRoutes can be defined. If empty, roots can be defined in any
-	// namespace.
-	IngressRouteRootNamespaces []string
-
-	mu sync.RWMutex
-
-	ingresses     map[meta]*v1beta1.Ingress
-	ingressroutes map[meta]*ingressroutev1.IngressRoute
-	secrets       map[meta]*v1.Secret
-	delegations   map[meta]*ingressroutev1.TLSCertificateDelegation
-	services      map[meta]*v1.Service
-}
-
-// meta holds the name and namespace of a Kubernetes object.
-type meta struct {
-	name, namespace string
-}
 
 const (
 	StatusValid    = "valid"
 	StatusInvalid  = "invalid"
 	StatusOrphaned = "orphaned"
 )
-
-// Insert inserts obj into the KubernetesCache.
-// If an object with a matching type, name, and namespace exists, it will be overwritten.
-func (kc *KubernetesCache) Insert(obj interface{}) {
-	kc.mu.Lock()
-	defer kc.mu.Unlock()
-	switch obj := obj.(type) {
-	case *v1.Secret:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		if kc.secrets == nil {
-			kc.secrets = make(map[meta]*v1.Secret)
-		}
-		kc.secrets[m] = obj
-	case *v1.Service:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		if kc.services == nil {
-			kc.services = make(map[meta]*v1.Service)
-		}
-		kc.services[m] = obj
-	case *v1beta1.Ingress:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		if kc.ingresses == nil {
-			kc.ingresses = make(map[meta]*v1beta1.Ingress)
-		}
-		kc.ingresses[m] = obj
-	case *ingressroutev1.IngressRoute:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		if kc.ingressroutes == nil {
-			kc.ingressroutes = make(map[meta]*ingressroutev1.IngressRoute)
-		}
-		kc.ingressroutes[m] = obj
-	case *ingressroutev1.TLSCertificateDelegation:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		if kc.delegations == nil {
-			kc.delegations = make(map[meta]*ingressroutev1.TLSCertificateDelegation)
-		}
-		kc.delegations[m] = obj
-
-	default:
-		// not an interesting object
-	}
-}
-
-// Remove removes obj from the KubernetesCache.
-// If no object with a matching type, name, and namespace exists in the DAG, no action is taken.
-func (kc *KubernetesCache) Remove(obj interface{}) {
-	switch obj := obj.(type) {
-	default:
-		kc.remove(obj)
-	case cache.DeletedFinalStateUnknown:
-		kc.Remove(obj.Obj) // recurse into ourselves with the tombstoned value
-	}
-}
-
-func (kc *KubernetesCache) remove(obj interface{}) {
-	kc.mu.Lock()
-	defer kc.mu.Unlock()
-	switch obj := obj.(type) {
-	case *v1.Secret:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		delete(kc.secrets, m)
-	case *v1.Service:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		delete(kc.services, m)
-	case *v1beta1.Ingress:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		delete(kc.ingresses, m)
-	case *ingressroutev1.IngressRoute:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		delete(kc.ingressroutes, m)
-	case *ingressroutev1.TLSCertificateDelegation:
-		m := meta{name: obj.Name, namespace: obj.Namespace}
-		delete(kc.delegations, m)
-	default:
-		// not interesting
-	}
-}
 
 // A Builder builds a *DAGs
 type Builder struct {
